+++ CHANGELOG.md
@@ -4,6 +4,10 @@
   session cookie given a valid ID token.
 - [added] A new `verify_session_cookie()` method for verifying a given
   cookie string is valid.
+- [added] Added the `mutable_content` optional field to the `messaging.Aps`
+  type.
+- [added] Added support for specifying arbitrary custom key-value
+  fields in the `messaging.Aps` type.
 
 # v2.9.1